+++ manager-application/src/utils/array.ts
