+++ manager-application/src/utils/dialog-data.ts
