+++ manager-application/src/utils/send-button-data.ts
